requirements.txt
protobuf
# Core dependencies
openai>=1.52.0
google-generativeai>=0.3.0
langsmith>=0.1.0
wandb>=0.16.0
python-dotenv>=1.0.0

# Data processing
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
rich>=13.7.0
tabulate>=0.9.0
